+++ src/main.rs
@@ -6,13 +6,16 @@
 
 use axum::{
     extract::{Extension},
+    body::{Body, Bytes},
     routing::{get, post},
     Router,
     extract::Request,
+    http::{response, StatusCode},
     response::{IntoResponse, Response},
-    middleware::{self, Next, map_response,},
+    middleware::{self, Next},
 };
-//use hyper::header::HeaderValue;
+use hyper::header::HeaderValue;
+use http_body_util::BodyExt;
 use std::env;
 use std::sync::{Arc, RwLock};
 use tower_http::trace::TraceLayer;
@@ -103,22 +106,22 @@ async fn main() {
             ServiceBuilder::new()
                 .layer(TraceLayer::new_for_http())
                 .into_inner(),
-        );
-        //.layer(map_response(set_header));// Add the shared config to the application state;
+        )
+        .layer(middleware::from_fn(response_version_header));
 
     // run it
     let listener = tokio::net::TcpListener::bind(&addr).await.unwrap();
     println!("listening on {}", listener.local_addr().unwrap());
     axum::serve(listener, app).await.unwrap();
 }
-// async fn set_header<B>(mut response: Response<B>, Extension(shared_state): Extension<Arc<AppState>>) -> Response<B> {
-//     let version = get_cargo_version().await.unwrap();
+async fn response_version_header(
+    request: Request, 
+    next: Next
+) -> Response {
+    let mut response = next.run(request).await;
 
-//     // Access the env_config values
-//     let env_config = shared_state.config.read().unwrap();
-//     let port = &env_config.env_file.port;
+    // do something with `response`...
+    response.headers_mut().insert("x-source-board-version", HeaderValue::from_static("1.0.0"));
 
-//     response.headers_mut().insert("x-source-board-version", HeaderValue::from_str(&version).unwrap());
-//     response.headers_mut().insert("x-source-board-port", HeaderValue::from_str(port).unwrap());
-//     response
-// }
+    response
+}