+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * .route("/github/user/rate-limit", get(github_get_user_rate_limit_handler))
 * Debug output JSON to file
+* Stage resource, stage deploy action
 
 ## 0.4.1